+++ src/SeablastPdo.php
@@ -70,6 +70,16 @@ class SeablastPdo extends PDO
     }
 
     /**
+     * PDOException may not be caught by SeablastPdo, so index.php can trigger it directly.
+     *
+     * @return void
+     */
+    public function indicateDatabaseError(): void
+    {
+            $this->databaseError = true;
+    }
+
+    /**
      * Determines if the query is read-only.
      *
      * @param string $query