+++ CHANGELOG.md
@@ -13,6 +13,8 @@ feat: add `ADMIN_TABLE_DELETE_ROW` and `ADMIN_TABLE_INSERT_ROW` permissions
 
 ### `Changed` for changes in existing functionality
 
+feat: refer to jquery/3.7.1 in BlueprintWeb.latte always
+
 ### `Deprecated` for soon-to-be removed features
 
 ### `Removed` for now removed features